Move speaker_id and microphone mocks to _mocks
